User Class passed all Postman Tests

+++ src/main/java/ru/practicum/shareit/item/Service/ItemService.java
@@ -1,21 +1,21 @@
 package ru.practicum.shareit.item.Service;
 
-import ru.practicum.shareit.item.BadRequestException;
+import ru.practicum.shareit.exceptions.BadRequestException;
 import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.model.Item;
 
 import java.util.List;
 
 public interface ItemService {
-    List<Item> getItems(int userId);
+    List<Item> getItems(Long userId);
 
-    ItemDto getItemById(int id);
+    ItemDto getItemById(Long id);
 
     List<Item> getItemsText(String text);
 
-    ItemDto createItem(int userId, ItemDto itemDto) throws BadRequestException;
+    ItemDto createItem(Long userId, ItemDto itemDto) throws BadRequestException;
 
-    ItemDto updateItemById(int userId, int id, ItemDto item) throws CloneNotSupportedException, BadRequestException;
+    ItemDto updateItemById(Long userId, Long id, ItemDto item) throws CloneNotSupportedException, BadRequestException;
 
     void deleteItemById(int id);
 }
